Relocate search jobs to generic job package.
Will allow extension in future and makes jobs explicit as a domain concept.
