+++ src/main/java/uk/gov/mca/beacons/api/jobs/steps/BeaconBatchJobProcessor.java
@@ -1,4 +1,4 @@
-package uk.gov.mca.beacons.api.search.jobs;
+package uk.gov.mca.beacons.api.jobs.steps;
 
 import java.util.List;
 import org.springframework.batch.item.ItemProcessor;
