+++ src/test/java/uk/gov/mca/beacons/api/jobs/configuration/BeaconBatchJobConfigurationIntegrationTest.java
@@ -1,13 +1,15 @@
-package uk.gov.mca.beacons.api.search.jobs;
+package uk.gov.mca.beacons.api.jobs.configuration;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 
 import java.util.ArrayList;
 import java.util.List;
+import org.junit.Before;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.batch.core.ExitStatus;
+import org.springframework.batch.core.Job;
 import org.springframework.batch.core.JobExecution;
 import org.springframework.batch.core.JobParameters;
 import org.springframework.batch.test.JobLauncherTestUtils;
@@ -33,9 +35,17 @@ public class BeaconBatchJobConfigurationIntegrationTest
   @Autowired
   BeaconSearchRepository beaconSearchRepository;
 
+  @Autowired
+  Job job;
+
   @AfterEach
   public void cleanUp() {
     jobRepositoryTestUtils.removeJobExecutions();
+  }
+
+  @Before
+  public void setUp() {
+    jobLauncherTestUtils.setJob(job);
   }
 
   @Test
